requirements.txt
pillow
flask 